+++ Common/src/main/resources/META-INF/common-plugin.xml
@@ -107,9 +107,10 @@
         <applicationService serviceImplementation="com.tabnineCommon.lifecycle.BinaryStateService"/>
         <applicationService serviceImplementation="com.tabnineCommon.capabilities.CapabilitiesService"/>
         <applicationService serviceImplementation="com.tabnineCommon.lifecycle.LifeCycleHelper"/>
-        <codeInsight.inlayProviderFactory implementation="com.tabnineCommon.chat.lens.TabnineLensProviderFactory" />
+        <codeInsight.inlayProvider language="JAVA" implementationClass="com.tabnineCommon.chat.lens.TabnineLensJavaProvider" />
+        <codeInsight.inlayProvider language="JavaScript" implementationClass="com.tabnineCommon.chat.lens.TabnineLensJavascriptProvider" />
+        <codeInsight.inlayProvider language="Python" implementationClass="com.tabnineCommon.chat.lens.TabnineLensPythonProvider" />
 
-        <actionPromoter implementation="com.tabnineCommon.inline.InlineActionsPromoter"/>
         <editorActionHandler action="EditorEscape" implementationClass="com.tabnineCommon.inline.EscapeHandler"
                              id="previewEscape" order="before hide-hints"/>
         <editorFactoryDocumentListener implementation="com.tabnineCommon.inline.TabnineDocumentListener"/>
